feat: Add a call via the function
closes: #86
