+++ src/hooks/loader/index.ts
@@ -1,4 +1,5 @@
 import useLoader from './loader.hook';
+import fnLoader from './loader.functions';
 
-export { useLoader };
+export { useLoader, fnLoader };
 export type { LoaderI } from './loader.types';
