+++ src/hooks/toast/index.ts
@@ -1,2 +1,5 @@
-export { useToast, toastSettings } from './toast.hook';
+import useToast from './toast.hook';
+
+export { fnToast, toastSettings } from './toast.functions';
 export type { ToastI, ToastPropsT, ToastSettingsI, ToastTypes, ToastDataObject, RecursivePartial } from './toast.types';
+export { useToast };
